refactor: rename encodeable to encodable
