+++ packages/superset-ui-encodable/src/parsers/createGetterFromChannelDef.ts
