+++ packages/superset-ui-encodable/src/parsers/format/createFormatterFromFieldTypeAndFormat.ts
@@ -1,7 +1,7 @@
 import { getNumberFormatter } from '@superset-ui/number-format';
 import { getTimeFormatter } from '@superset-ui/time-format';
-import { Type } from '../types/VegaLite';
-import { Formatter } from '../types/ChannelDef';
+import { Type } from '../../types/VegaLite';
+import { Formatter } from '../../types/ChannelDef';
 import fallbackFormatter from './fallbackFormatter';
 
 export default function createFormatterFromFieldTypeAndFormat(
