+++ packages/superset-ui-encodable/src/parsers/scale/applyAlign.ts
