+++ packages/superset-ui-encodable/src/parsers/scale/applyClamp.ts
