+++ packages/superset-ui-encodable/src/parsers/scale/applyDomain.ts
