+++ packages/superset-ui-encodable/src/parsers/scale/applyInterpolate.ts
