+++ packages/superset-ui-encodable/src/parsers/scale/applyNice.ts
