+++ packages/superset-ui-encodable/src/parsers/scale/applyRange.ts
