+++ packages/superset-ui-encodable/src/parsers/scale/applyZero.ts
