+++ packages/superset-ui-encodable/src/parsers/scale/createScaleFromScaleConfig.ts
