+++ packages/superset-ui-encodable/src/parsers/scale/createScaleFromScaleType.ts
