+++ packages/superset-ui-encodable/src/parsers/scale/getScaleCategoryFromScaleType.ts
