+++ packages/superset-ui-encodable/src/parsers/scale/inferScaleType.ts
