+++ packages/superset-ui-encodable/src/parsers/scale/isPropertySupportedByScaleType.ts
