+++ packages/superset-ui-encodable/src/parsers/scale/scaleCategories.ts
