+++ packages/superset-ui-encodable/src/typeGuards/Base.ts
