+++ packages/superset-ui-encodable/src/typeGuards/Scale.ts
