+++ packages/superset-ui-encodable/src/types/ChannelDef.ts
