+++ packages/superset-ui-encodable/src/types/Legend.ts
