+++ packages/superset-ui-encodable/src/utils/inferElementTypeFromUnionOfArrayTypes.ts
