+++ packages/superset-ui-encodable/src/utils/isEnabled.ts
