+++ packages/superset-ui-encodable/test/parsers/scale/getScaleCategoryFromScaleType.test.ts
