+++ packages/superset-ui-encodable/test/parsers/scale/inferScaleType.test.ts
