+++ packages/superset-ui-encodable/test/parsers/scale/isPropertySupportedByScaleType.test.ts
