+++ packages/superset-ui-encodable/test/typeGuards/Scale.test.ts
