--- packages/superset-ui-encodeable/README.md
@@ -1,23 +1,0 @@
-## @superset-ui/encodeable
-
-[![Version](https://img.shields.io/npm/v/@superset-ui/encodeable.svg?style=flat)](https://img.shields.io/npm/v/@superset-ui/encodeable.svg?style=flat)
-[![David (path)](https://img.shields.io/david/apache-superset/superset-ui.svg?path=packages%2Fsuperset-ui-encodeable&style=flat-square)](https://david-dm.org/apache-superset/superset-ui?path=packages/superset-ui-encodeable)
-
-Description
-
-#### Example usage
-
-```js
-import { xxx } from '@superset-ui/encodeable';
-```
-
-#### API
-
-`fn(args)`
-
-- Do something
-
-### Development
-
-`@data-ui/build-config` is used to manage the build configuration for this package including babel
-builds, jest testing, eslint, and prettier.
